initial.rev3
Instead of putting canvas into tab i'm putting into DOMWin, as my end
goal is to put into a panel anyways and not into a tab. Also I updated
the ctypes so it takes a section of the screen instead of just from the
top left, well I think I did, I cant get canvas to draw so I dont know
the contents yet, but the RGB looks right.