1.0-night.rev42
Fixed a bunch of issues with GTK as tested on multi monitor setup on
Ubuntu today. If go fullscreen it does cover everything but only on the
single monitor. This also lands making the drawing go over the other
fullscreen windows. However its not consistent. This also is not
covering the special UI areas. But its now spanning across all monitors.