1.1-night.rev29
The memory leak was due to consol.log in bootstrap.js was getting colMon
stuck in _consoleStorage. So on production release I'll remove those, so
it won't have the memleak. So for now restored the console.log's and
fixed two areas which were succeptible to error due to console.log being
on same line as brackets needed for other code.