1.1-night.rev41
Made so when no sliphover due to small window, then its link clickable
on images.

Also made view button link clickable if any of hte modifier keys are
down, and also right click to get to link location works on it.

Wass missing woff2 file for v4.3.0 so brought it in

Handling broken imges, before it would be blank, now it force shows that
borken image icon, nice

Functionalized save-* delete. I made it throw when it can't find the
image (when its broken) because it didnt delete anything, maybe user
moved it to somewhere else. So in that case the user should "remove from
history".

Fixed so when nothing left in gFileJson, it doesn't write a comma, as
that will lead to multiple commas at the start which will be an issue

Functionalized Imgure delete too