1.1-night.rev48
Finally!!! Oh my gosh got magnific working in xhtml, issues was the
multiply character used by magnific, i opened an issue:
https://github.com/dimsemenov/Magnific-Popup/issues/734

Shift clicking currently is getting hijacked by magfnific. I have to
find a way to block that. Because shift+click on the delegate will open
it in a new window when magfnific doesnt hijack.

Path to bootstrap-slider.js was bad, fixed that.