1.1-night.rev53
Added in the isotope no results message.

It seems that isotope.remove is not removing the items from the array,
this is really weird, and is causing some funky behavior, i need to
investigate this more, but its not so bad it only messes with the
showing of the no results message. so can revisit in future version. so
even though there are no items (as they were removed) the
filteredItems.length is > 0 so it doesnt show the no rez message, this
is quite weird, somewhat edge case though, so future revisit.