1.2-night.rev2
Switched to mozFetchAsStream, which is better because i dont have to get
blob then read to buffer and I can directly append it to mpis. With
buffer i had to set it to arraybuffer stream, and besides I couldnt
figure out why but it would be corrupt for bigger imgaes.
mozFetchAsStream is much more straight forward and is not causing any
corruptions on bigger images.