1.6b.rev27
* Made the browse file picker setLevel over the top of canvas which is
always on top, this is better then old method which was "set canvas to
normal, then back to top, after sync picker" because i was setting it to
normal with NSUInteger 0 and if normal then i would think the dock and
top menu strip should show. but it didnt. so it seems i was relying on a
quirk.