alt to gRefWidth - it was setState on mousemove
I CAN UNDO THIS commit if i want. because this perf is not as good as as
the gRefWidth technique BUT this technique makes the pal draggable and
all mousemove realted stuff muuuuuch smoother, its usable without really
noticing the lag. MAYBE use gRefWidth technique just for the width and
height part and onmouseup update react set state - HOWEVER do this only
after I try using shouldComponentUpdate for perf