trying svg to icon - not working
this branch created for issue #33 - i am trying to use @noureddin
awesome technique of svg and internal stylesheet. i got the stylesheet
overriding per os and it works when access raw, but when used as
list-image-style it wont work and thats how the CustomizableUI.jsm makes
the the toolbarbutton :( - need to figure out how to get the internal
style sheet to work with list-style-image otherwise we'll have to make
unique image per, which will stick. the unique styleshet per is much
better